[Backport] Security bug 905509 (6/13)
Manual backport of original patch by Antoine Labour
<piman@chromium.org>:
Make ClientTransferCache entries use a uint32_t size

As they go into transfer buffers, they need to fit within 4GB, so check that at
the source. Some places were silently clamping size_t to uint32_t, which could
be a theoretical issue.

Bug: 905509
Reviewed-on: https://chromium-review.googlesource.com/c/1400046

+++ chromium/cc/paint/image_transfer_cache_entry.h
@@ -31,7 +31,7 @@ class CC_PAINT_EXPORT ClientImageTransferCacheEntry
   uint32_t Id() const final;
 
   // ClientTransferCacheEntry implementation:
-  size_t SerializedSize() const final;
+  uint32_t SerializedSize() const final;
   bool Serialize(base::span<uint8_t> data) const final;
 
  private:
@@ -67,7 +67,7 @@ class CC_PAINT_EXPORT ServiceImageTransferCacheEntry
   GrContext* context_;
   sk_sp<SkImage> image_;
   bool has_mips_ = false;
-  size_t size_ = 0;
+  uint32_t size_ = 0;
   bool fits_on_gpu_ = false;
 };
